Move the creation of the client to a separate file, create_client.py
All of ingest.py, query.py, and delete_client.py contained the same code for creating
a client to interact with the OpenSearch instance. This commit moves that code out of those
files and into a separate file.
