+++ query.py
@@ -1,31 +1,6 @@
-from opensearchpy import OpenSearch, RequestsHttpConnection, AWSV4SignerAuth
-import os
-from dotenv import load_dotenv
-import certifi
-import boto3
+from create_client import create_client
 
-load_dotenv()
-
-
-host = os.getenv('OPENSEARCH_HOST')
-port = os.getenv('OPENSEARCH_PORT')
-region = 'us-east-1'
-
-service = 'aoss'
-credentials = boto3.Session().get_credentials()
-auth = AWSV4SignerAuth(credentials, region, service)
-
-
-# Create the client with SSL/TLS enabled, but hostname verification disabled.
-client = OpenSearch(
-    hosts=[{'host': host, 'port': port}],
-    http_compress = True, # enables gzip compression for request bodies
-    http_auth=auth,
-    use_ssl=True,
-    verify_certs=True,
-    connection_class=RequestsHttpConnection,
-    pool_maxsize=20,
-)
+client = create_client()
 
 index_name = "comments"
